requirements.txt
pandas
matplotlib
seaborn
langchain
google-generativeai
langchain-google-genai
langchain-experimental
python-dotenv